requirements.txt
discord.py==2.3.2
python-dotenv==1.0.0
google-api-python-client==2.100.0
aiohttp>=3.9.5
asyncpg>=0.29.0